Rename helpers.ts and move to `helpers` directory
