+++ __test__/unit/pushSubscription/nativePermissionChange.test.ts
@@ -5,7 +5,7 @@ import PermissionManager from "../../../src/shared/managers/PermissionManager";
 import EventHelper from "../../../src/shared/helpers/EventHelper";
 import { SubscriptionManager } from "../../../src/shared/managers/SubscriptionManager";
 import { DUMMY_PUSH_TOKEN } from "../../support/constants";
-import { initializeWithPermission } from "./helpers";
+import { initializeWithPermission } from "../../support/helpers/pushSubscription";
 
 describe('Notification Types are set correctly on subscription change', () => {
   beforeEach(async () => {
